Split deploy tests into 1 file per view.
